+++ lib/helper/news.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:news_app/models/article_model.dart';
+import 'package:news_app/models/article.dart';
 import 'package:http/http.dart' as http;
 class News{
   List<ArticleModel> news = [];
