+++ lib/models/article.dart
