+++ lib/views/category_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:news_app/models/article_model.dart';
+import 'package:news_app/models/article.dart';
 import 'package:news_app/views/article_view.dart';
 import 'package:news_app/helper/news.dart';
