+++ lib/views/home.dart
@@ -2,8 +2,8 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:news_app/helper/data.dart';
-import 'package:news_app/models/article_model.dart';
-import 'package:news_app/models/category_model.dart';
+import 'package:news_app/models/article.dart';
+import 'package:news_app/models/category.dart';
 import 'package:news_app/helper/news.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:news_app/views/article_view.dart';
